MODNCIP-43: Support holdings-storage 6.0 interface
mod-inventory-storage has new major interface versions:

instance-storage 9.0
holdings-storage 6.0
item-storage 10.0

mod-ncip only uses the holdings-storage interfaces, it doesn't use the two other interfaces.

mod-ncip is not affected because the incompatible change is in the DELETE all APIs only that is not used by mod-ncip.

Only the okapiInterfaces need to be bumped in "requires" section of ModuleDescriptor-template.json